+++ crypto/crypto_small/sha512_small.h
@@ -1,8 +1,6 @@
 ﻿#ifndef _SHA512_SMALL_H_
 #define _SHA512_SMALL_H_
 
-#include <memory.h>
-
 typedef struct _sha512_ctx {
 	unsigned __int64 hash[8];
 	unsigned __int64 length;
@@ -15,7 +13,7 @@ typedef struct _sha512_ctx {
 #define SHA512_BLOCK_SIZE  128
 
 void sha512_init(sha512_ctx *ctx);
-void sha512_hash(sha512_ctx *ctx, const unsigned char *in, size_t inlen);
+void sha512_hash(sha512_ctx *ctx, const unsigned char *in, unsigned long inlen);
 void sha512_done(sha512_ctx *ctx, unsigned char *out);
 
 #endif
