+++ crypto/crypto_small/xts_small_aes.h
@@ -1,7 +1,6 @@
 ﻿#ifndef _XTS_SMALL_AES_H_
 #define _XTS_SMALL_AES_H_
 
-#include <memory.h>
 #include "aes_small.h"
 #include "aes_padlock_small.h"
 
@@ -22,12 +21,12 @@ typedef __declspec(align(16)) struct _xts_key {
 } xts_key;
 
 void xts_aes_set_key(const unsigned char *key, int alg, xts_key *skey);
-void xts_aes_encrypt(const unsigned char *in, unsigned char *out, size_t len, unsigned __int64 offset, xts_key *key);
-void xts_aes_decrypt(const unsigned char *in, unsigned char *out, size_t len, unsigned __int64 offset, xts_key *key);
+void xts_aes_encrypt(const unsigned char *in, unsigned char *out, unsigned long len, unsigned __int64 offset, xts_key *key);
+void xts_aes_decrypt(const unsigned char *in, unsigned char *out, unsigned long len, unsigned __int64 offset, xts_key *key);
 void xts_aes_init(int hw_crypt);
 
 typedef void (*xts_setkey_proc)(const unsigned char *key, int alg, xts_key *skey);
-typedef void (*xts_crypt_proc) (const unsigned char *in, unsigned char *out, size_t len, unsigned __int64 offset, xts_key *key);
+typedef void (*xts_crypt_proc) (const unsigned char *in, unsigned char *out, unsigned long len, unsigned __int64 offset, xts_key *key);
 typedef void (*xts_init_proc)  (int hw_crypt);
 
 #define xts_set_key xts_aes_set_key
